Feat/#344: 수정 ShopView 하위뷰 이름변경
