+++ MC3Team18/MC3Team18/View/ShopView.swift
@@ -53,7 +53,7 @@ struct ShopView: View {
     @State var shopItem: [ShopItem] = []
     @State var selectedItem: ShopItem?
     @State var tappedItem: ShopItem = ShopItem(itemName: "", itemNameKorean: "", price: 0, itemCategory: .acc, itemStatus: 0)
-    @State var isPurchasePopupAppear: Bool = false
+    @State var showPurchasePopup: Bool = false
     @State var buyable: Bool = false
     var body: some View {
         ZStack{
@@ -62,16 +62,16 @@ struct ShopView: View {
             VStack (spacing: 10){
                 ShopCharacterView(selectedItem: $selectedItem)
                 
-                ShopAccessoriesScrollView(shopItem: $shopItem, selectedItem: $selectedItem, tappedItem: $tappedItem, isPurchasePopupAppear: $isPurchasePopupAppear, buyable: $buyable)
+                ShopItemScrollView(shopItem: $shopItem, selectedItem: $selectedItem, tappedItem: $tappedItem, isPurchasePopupAppear: $showPurchasePopup, buyable: $buyable)
             }
             .padding(.top, 20)
             
-            if isPurchasePopupAppear {
+            if showPurchasePopup {
                 Color.Black.opacity(0.8).ignoresSafeArea()
                     .onTapGesture {
-                        isPurchasePopupAppear = false
+                        showPurchasePopup = false
                     }
-                ShopItemPurchaseView(shopItem: $shopItem, isPurchasePopupAppear: $isPurchasePopupAppear, tappedItem: $tappedItem, buyAble: buyable)
+                ShopPopupView(shopItem: $shopItem, showPurchasePopup: $showPurchasePopup, tappedItem: $tappedItem, buyAble: buyable)
             }
         }
         .onAppear {
